chore: change components directory
