+++ pages/_app.tsx
@@ -1,9 +1,9 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
-import NavBar from "./_component/pages/NavBar";
+import NavBar from "./_component/pages/global/NavBar";
 import { RecoilRoot } from "recoil";
 import { roboto } from "@/util/font";
-import MobileMenu from "./_component/pages/MobileMenu";
+import MobileMenu from "./_component/pages/global/MobileMenu";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
