+++ util/constant.ts
@@ -1,6 +1,6 @@
 export const shadow = "0 2px 4px #00000010, 0 1px 2px #00000010";
 
-export const mainColor = "#0D689E";
+export const mainColor = "#0271a1";
 export const subColor1 = "#84B4D4";
 export const subColor2 = "#BAD5E8";
 export const gray = "#F1F5F9";
